Travail sur scripts installation
Detection packages deja installes

+++ bin/install_python.sh
@@ -12,16 +12,27 @@ python3 -m venv --system-site-packages $PATH_VENV
 echo "Activer venv ${PATH_VENV}"
 source "${PATH_VENV}/bin/activate"
 
-echo "Installer pip wheel"
-pip3 install wheel
+if ! pip3 list | grep "wheel" > /dev/null; then
+  echo "[INFO] Installer pip wheel"
+  pip3 install wheel
+fi
 
-echo "Installer millegrilles messages (path $URL_MGMESSAGES)"
-pip3 install $URL_MGMESSAGES
+# Verifier que le package millegrilles-messages de la bonne version est installe
+MGMESSAGES_INSTALLE=0
+if ! pip3 list | grep -e "${MG_PIP_PACKAGE_NAME}" | grep -e "${MG_PIP_PACKAGE_VERSION}" > /dev/null; then
+  echo "[INFO] Installer millegrilles messages (path $MG_PIP_PACKAGE_URL)"
+  pip3 install $URL_MGMESSAGES
+  MGMESSAGES_INSTALLE=1
+fi
 
+echo "[INFO] Verifier requirements python pour millegrilles, installer au besoin"
 pip3 install -r requirements.txt
 
 # Fix pour cryptography
 # https://stackoverflow.com/questions/74981558/error-updating-python3-pip-attributeerror-module-lib-has-no-attribute-openss
-python3 -m pip install -U pyOpenSSL cryptography
+if [ "${MGMESSAGES_INSTALLE}" -eq 1 ]; then
+  echo "[INFO] Fix installation cryptography"
+  python3 -m pip install -U pyOpenSSL cryptography
+fi
 
 echo "[INFO] Fin configuration venv python3"
